requirements.txt
# Backend
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.4.2

# ML & Data
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
xgboost==2.0.0
joblib==1.3.2

# Frontend & API Calls
streamlit>=1.32.0
plotly==5.17.0
requests==2.31.0

# Notebook & Reporting
notebook==7.0.6
openpyxl==3.1.2

# For loading environment variables (.env file)
python-dotenv==1.0.0